+++ src/index.ts
@@ -8,26 +8,84 @@ import {
 } from './Overlay';
 
 class OverlayLayer extends Overlay {
+  private element: HTMLElement | undefined;
+  private _size: Size | undefined;
   constructor(options: object) {
-    super(options);
+    super(Object.assign(options, {
+      stopEvent: false, // must be false
+      offset: [0, 0],
+    }));
+
+    /**
+     * size
+     */
+    this._size = undefined;
+  }
+
+  setMap (map: any): void {
+    super.setMap(map);
+  }
+
+  /**
+   * overwrite setOffset: must be [0, 0]
+   * @param offset
+   */
+  setOffset(offset: number[] = [0, 0]) {
+    super.setOffset(offset && [0, 0]);
+  }
+
+  /**
+   * overwrite setPositioning: must be top-left
+   * @param positioning
+   */
+  setPositioning(positioning: OverlayPositioning = 'top-left') {
+    super.setPositioning(positioning && 'top-left');
   }
 
   panIntoView(){}
 
   /**
+   * bind map update
+   */
+  render(): void {
+    this.dispatchEvent({
+      type: 'preupdate',
+      value: this,
+    });
+    this.updatePixelPosition();
+  }
+
+  /**
    * overwrite getPosition method: top-left;
    */
-  getPosition(): Coordinate {
+  getPosition(): Coordinate | void {
     const map = this.getMap();
+    if (!map) return;
     const size = map.getSize();
     const extent = map.getView().calculateExtent(size);
     return [extent[0], extent[3]];
+  }
+
+  /**
+   * update view size
+   * @param size
+   */
+  updateViewSize(size: Size): void {
+    if (!this.element) return;
+    this._size = size;
+    this.element.style.width = `${size[0]}px`;
+    this.element.style.height = `${size[1]}px`;
+    this.element.setAttribute('width', String(size[0]));
+    this.element.setAttribute('height', String(size[1]));
   }
 
   getPositioning(): OverlayPositioning {
     return 'top-left';
   }
 
+  /**
+   * update layer position, must be [0, 0]
+   */
   updatePixelPosition(){
     const map = this.getMap();
     const position = this.getPosition();
@@ -36,70 +94,21 @@ class OverlayLayer extends Overlay {
       return;
     }
 
-    const pixel = map.getPixelFromCoordinate(position);
+    // const pixel = map.getPixelFromCoordinate(position);
     const mapSize = map.getSize();
-    this.updateRenderedPosition(pixel, mapSize);
+    if (!this._size) {
+      this.updateViewSize(mapSize);
+    } else {
+      const _update = this._size[0] === 0 || this._size[1] === 0
+        || (this._size[0] !== mapSize[0] && this._size[1] !== mapSize[1]);
+      _update && this.updateViewSize(mapSize);
+    }
+    this.updateRenderedPosition([0, 0], mapSize);
+    this.dispatchEvent({
+      type: 'updated',
+      value: position,
+    });
   }
-
-  // updateRenderedPosition (pixel, mapSize) {
-  //   var style = this.element.style;
-  //   var offset = this.getOffset();
-  //
-  //   var positioning = this.getPositioning();
-  //
-  //   this.setVisible(true);
-  //
-  //   var offsetX = offset[0];
-  //   var offsetY = offset[1];
-  //   if (positioning == OverlayPositioning.BOTTOM_RIGHT ||
-  //     positioning == OverlayPositioning.CENTER_RIGHT ||
-  //     positioning == OverlayPositioning.TOP_RIGHT) {
-  //     if (this.rendered.left_ !== '') {
-  //       this.rendered.left_ = style.left = '';
-  //     }
-  //     var right = Math.round(mapSize[0] - pixel[0] - offsetX) + 'px';
-  //     if (this.rendered.right_ != right) {
-  //       this.rendered.right_ = style.right = right;
-  //     }
-  //   } else {
-  //     if (this.rendered.right_ !== '') {
-  //       this.rendered.right_ = style.right = '';
-  //     }
-  //     if (positioning == OverlayPositioning.BOTTOM_CENTER ||
-  //       positioning == OverlayPositioning.CENTER_CENTER ||
-  //       positioning == OverlayPositioning.TOP_CENTER) {
-  //       offsetX -= this.element.offsetWidth / 2;
-  //     }
-  //     var left = Math.round(pixel[0] + offsetX) + 'px';
-  //     if (this.rendered.left_ != left) {
-  //       this.rendered.left_ = style.left = left;
-  //     }
-  //   }
-  //   if (positioning == OverlayPositioning.BOTTOM_LEFT ||
-  //     positioning == OverlayPositioning.BOTTOM_CENTER ||
-  //     positioning == OverlayPositioning.BOTTOM_RIGHT) {
-  //     if (this.rendered.top_ !== '') {
-  //       this.rendered.top_ = style.top = '';
-  //     }
-  //     var bottom = Math.round(mapSize[1] - pixel[1] - offsetY) + 'px';
-  //     if (this.rendered.bottom_ != bottom) {
-  //       this.rendered.bottom_ = style.bottom = bottom;
-  //     }
-  //   } else {
-  //     if (this.rendered.bottom_ !== '') {
-  //       this.rendered.bottom_ = style.bottom = '';
-  //     }
-  //     if (positioning == OverlayPositioning.CENTER_LEFT ||
-  //       positioning == OverlayPositioning.CENTER_CENTER ||
-  //       positioning == OverlayPositioning.CENTER_RIGHT) {
-  //       offsetY -= this.element.offsetHeight / 2;
-  //     }
-  //     var top = Math.round(pixel[1] + offsetY) + 'px';
-  //     if (this.rendered.top_ != top) {
-  //       this.rendered.top_ = style.top = top;
-  //     }
-  //   }
-  // };
 
   getMap() {
     return super.getMap();
@@ -107,6 +116,15 @@ class OverlayLayer extends Overlay {
 
   setVisible(visible: Boolean): void {
     super.setVisible(visible);
+    this.dispatchEvent({
+      type: 'change:visible ',
+      value: visible,
+    });
+  }
+
+  // @ts-ignore
+  dispatchEvent(...args) {
+    super.dispatchEvent(...args);
   }
 
   updateRenderedPosition(pixel: Pixel, mapSize: Size): void {